Split flood_simulation.rs into submodule directory (1196 -> ~300 lines each)
Reorganize flood_simulation.rs into flood_simulation/ directory with:
- damage_curves.rs: constants, depth-damage interpolation functions and tests (484 lines)
- resources.rs: FloodGrid and FloodState resources and tests (211 lines)
- systems.rs: update_flood_simulation system and FloodSimulationPlugin (235 lines)
- system_tests.rs: water spreading, drainage, and building damage tests (301 lines)
- mod.rs: re-exports all public items for backward compatibility (32 lines)

Part of #1389
